@@ -1,48 +1,30 @@
 {
 	"name": "frontend",
 	"version": "0.3.0",
-	"private": true,
 	"scripts": {
-		"dev": "sapper dev",
-		"build": "sapper build --legacy",
-		"export": "sapper export --legacy",
-		"start": "node __sapper__/build",
-		"validate": "svelte-check --ignore src/node_modules/@sapper"
-	},
-	"dependencies": {
-		"compression": "^1.7.1",
-		"express": "^4.17.1",
-		"polka": "next",
-		"postcss": "^8.3.6",
-		"sirv": "^1.0.0",
-		"svelte-markdown": "^0.1.12"
+		"dev": "svelte-kit dev",
+		"build": "svelte-kit build",
+		"preview": "svelte-kit preview",
+		"check": "svelte-check --tsconfig ./tsconfig.json",
+		"check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
+		"lint": "prettier --ignore-path .gitignore --check --plugin-search-dir=. .",
+		"format": "prettier --ignore-path .gitignore --write --plugin-search-dir=. ."
 	},
 	"devDependencies": {
-		"@babel/core": "^7.0.0",
-		"@babel/plugin-syntax-dynamic-import": "^7.0.0",
-		"@babel/plugin-transform-runtime": "^7.0.0",
-		"@babel/preset-env": "^7.0.0",
-		"@babel/runtime": "^7.0.0",
-		"@rollup/plugin-babel": "^5.0.0",
-		"@rollup/plugin-commonjs": "^14.0.0",
-		"@rollup/plugin-node-resolve": "^8.0.0",
-		"@rollup/plugin-replace": "^2.4.0",
-		"@rollup/plugin-typescript": "^6.0.0",
-		"@rollup/plugin-url": "^5.0.0",
-		"@tsconfig/svelte": "^1.0.10",
-		"@types/compression": "^1.7.0",
-		"@types/node": "^14.11.1",
-		"@types/polka": "^0.5.1",
-		"autoprefixer": "^10.3.4",
+		"@sveltejs/kit": "next",
+		"@types/node": "^16.9.1",
+		"prettier": "~2.2.1",
+		"prettier-plugin-svelte": "^2.2.0",
+		"svelte": "^3.34.0",
+		"svelte-check": "^2.0.0",
+		"svelte-preprocess": "^4.0.0",
+		"tslib": "^2.0.0",
+		"typescript": "^4.0.0"
+	},
+	"type": "module",
+	"dependencies": {
+		"@sveltejs/adapter-node": "^1.0.0-next.46",
 		"node-sass": "^6.0.1",
-		"rollup": "^2.3.4",
-		"rollup-plugin-svelte": "^7.0.0",
-		"rollup-plugin-terser": "^7.0.0",
-		"sapper": "^0.28.0",
-		"svelte": "^3.17.3",
-		"svelte-check": "^1.0.46",
-		"svelte-preprocess": "^4.9.4",
-		"tslib": "^2.0.1",
-		"typescript": "^4.0.3"
+		"svelte-markdown": "^0.1.12"
 	}
 }
