@@ -1,11 +1,19 @@
-const sveltePreprocess = require("svelte-preprocess");
-module.exports = {
-	preprocess: sveltePreprocess({
-		scss: {
-			includePaths: ["src"],
-		},
-		postcss: {
-			plugins: [require("autoprefixer")],
-		},
-	}),
+import preprocess from 'svelte-preprocess';
+import node from '@sveltejs/adapter-node';
+
+/** @type {import('@sveltejs/kit').Config} */
+const config = {
+	// Consult https://github.com/sveltejs/svelte-preprocess
+	// for more information about preprocessors
+	preprocess: preprocess(),
+
+	kit: {
+		// hydrate the <div id="svelte"> element in src/app.html
+		target: '#svelte',
+		adapter: node({
+			out: 'build'
+		})
+	}
 };
+
+export default config;
