@@ -1,12 +1,12 @@
 <script lang="ts">
 </script>
 
-<p class={"paragraph"} {...$$props}>
+<p class={'paragraph'} {...$$props}>
 	<slot />
 </p>
 
 <style lang="scss">
-	@import "../styles/includes";
+	@import '../../styles/includes';
 
 	.paragraph {
 		margin-block-start: 1em;
