@@ -10,5 +10,5 @@
 </main>
 
 <style lang="scss" global>
-	@import "../styles/global.scss";
+	@import '../styles/global.scss';
 </style>
